requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
asyncpg==0.27.0
attrs==22.1.0
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
discord==2.0.0
discord.py==2.1.0
frozenlist==1.3.3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
multidict==6.0.3
Pillow==9.3.0
psycopg2==2.9.5
schedule==1.1.0
wavelink==1.3.4
Werkzeug==2.2.2
yarl==1.8.2

genshin~=1.5.0